wip: conditional init
