+++ src/main.tsx
@@ -1,15 +1,11 @@
-import './instrument.ts';
-import './i18initializer';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import { setupLogger } from './utils/shared-logger.ts';
-import App from './App';
+import AppWrapper from './AppWrapper.tsx';
 
 const root = createRoot(document.getElementById('root') as HTMLElement);
-setupLogger();
 
 root.render(
     <StrictMode>
-        <App />
+        <AppWrapper />
     </StrictMode>
 );
